Mir: Splits up mir_elivra{2,3}d tests by policy to reduce compilation time/memory
